requirements.txt
NumPy version: 1.24.3
PyTorch version: 2.1.1+cu121
Scikit-learn version: 1.3.2
Pandas version: 2.0.2
Matplotlib version: 3.4.3